Убраны файлы для работы с wifi

+++ lib/grbl/src/serial.hpp
@@ -41,7 +41,7 @@ void serial_init();
 void serial_write(uint8_t data);
 
 // Fetches the first byte in the serial read buffer. Called by main program.
-uint8_t serial_read(uint8_t client);
+uint16_t serial_read(uint8_t client);
 
 // Reset and empty data in read buffer. Used by e-stop and reset.
 void serial_reset_read_buffer(uint8_t client);
